requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
decorator==4.1.2
google-api-core==0.1.2
google-auth==1.2.1
google-cloud-vision==0.29.0
googleapis-common-protos==1.5.3
grpcio==1.7.3
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
parso==0.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
